pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mellymell"
version = "0.1.0"
description = "Open-source pitch detection prototype (Python)"
readme = "README.md"
authors = [{ name = "David Ryan" }]
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.11",
  "sounddevice>=0.4",
  "soundfile>=0.12",
  "librosa>=0.10",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
ml = [
  "torch>=2.0",
  "torchcrepe>=0.0.20",
  "pesto-pitch>=0.1.0",
]
crepe = [
  "torch>=2.0",
  "torchcrepe>=0.0.20",
]
pesto = [
  "torch>=2.0",
  "pesto-pitch>=0.1.0",
]
polyphonic = [
  "basic-pitch>=0.3.0",
]
all = [
  "mellymell[ml,polyphonic,dev]",
]

[project.urls]
Homepage = "https://github.com/ddri/mellymell"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mellymell*"]

